Triggered Build: 24.01.004                                             Original message: Refactor version and app name retrieval methods
A new class 'Version' was created in the 'utils' module to handle the retrieval of the application's version and name from 'version.txt' and 'app_name.txt' respectively. The 'load_env' method in the 'configs' module has been updated to use this new class. This makes the method for retrieving these values more consistent and isolates it to a specific class.