Add method to convert dictionary keys to lowercase in utils
A static method 'conv_keys_to_lower' has been added to the 'utils' module to convert all dictionary keys to lowercase. This method is used in the 'interceptor' module especially to handle cases where specific header fields, like 'x-correlation-id', are not preset in request headers. This change aids in better handling and normalization of header data.